pyproject.toml
[project]
name = "opendeepresearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "ipykernel>=7.1.0",
    "jinja2>=3.1.6",
    "langchain>=1.0.5",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.3",
    "rich>=14.2.0",
    "streamlit>=1.52.1",
    "watchdog>=6.0.0",
]
[tool.setuptools.packages.find]


